Refactor middleware to a common location
